bid_scrapy_project/requirements.txt
anyio==3.7.0
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
Deprecated==1.2.14
exceptiongroup==1.1.1
filelock==3.12.2
gne==0.3.0
h11==0.14.0
h2==4.1.0
hpack==3.0.0
httpcore==0.17.2
hyperframe==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kafka-python==2.0.2
loguru==0.7.0
lxml==4.9.2
numpy==1.24.3
packaging==23.1
pandas==1.5.3
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyDispatcher==2.0.7
PyExecJS==1.5.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
redis==4.0.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.5.1
scrapy-redis==0.7.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tldextract==3.4.4
Twisted==22.4.0
twisted-iocpsupport==1.0.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
w3lib==2.1.1
win32-setctime==1.1.0
wrapt==1.15.0
zope.interface==6.0
